requirements.txt
camelot-py==0.10.1
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
deprecation==2.1.0
et-xmlfile==1.1.0
ghostscript==0.7
img2pdf==0.4.4
lxml==4.9.2
numpy==1.24.0
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pdfminer.six==20221105
pikepdf==6.2.6
Pillow==9.3.0
pycparser==2.21
pypdf2==2.12.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
tabulate==0.9.0
typing-extensions==4.4.0
